Changing the type of the Amount property.

+++ Models/Transaction.cs
@@ -5,7 +5,7 @@ namespace ApolloBank.Models
     public class Transaction
     {
         public int Id { get; private set; }
-        public int Amount { get; private set; }
+        public double Amount { get; private set; }
         public string? To { get; private set; }
         public string? From { get; private set; }
         public DateTime Date { get; private set; }
@@ -13,7 +13,7 @@ namespace ApolloBank.Models
         public TransactionType Transaction_Type { get; private set; }
         public int Account_Id { get; set; }
 
-        public Transaction(int id, int amount, string? to, string? from, DateTime date, string description, TransactionType transaction_Type, int account_Id)
+        public Transaction(int id, double amount, string? to, string? from, DateTime date, string description, TransactionType transaction_Type, int account_Id)
         {
             Id = id;
             Amount = amount;
